Import 1.3.62

+++ include/linux/sched.h
@@ -314,7 +314,7 @@ extern struct timeval xtime;
 extern int need_resched;
 extern void do_timer(struct pt_regs *);
 
-extern unsigned long * prof_buffer;
+extern unsigned int * prof_buffer;
 extern unsigned long prof_len;
 extern unsigned long prof_shift;
